in progress add LMTP client to James
Peer routing list divided into SMTP and LMTP hosts.
SMTP transport as usual.
LMTP transport: need a way to specify type (if it even exists)
The gateway field is only there because currently, it's the
only known way to specify port 24.  This will fail because
James sends EHLO to Dovecot instead of LHLO.
This webpage implies it is possible to override default type,
(see "gateway:port") but no corresponding documentation is found:
https://james.apache.org/server/3/apidocs/org/apache/james/transport/mailets/RemoteDelivery.html